Enhance role management and update project configuration
Updated `AdminDashboard.csproj` to prevent errors on duplicate publish output files. Improved `RoleController.cs` with formatting changes and added checks to prevent deletion of critical roles. Modified `Program.cs` to allow developer exception page in both environments and removed unnecessary middleware. Updated `Index.cshtml` to disable buttons for "Admin" and "User" roles.

Additionally, introduced new DTO classes in `AdminDto.cs`, `AppUser.cs`, `CategoryResponseDto.cs`, `CreateCategoryDto.cs`, and `LoginDTO.cs` to enhance data management and validation.

+++ AdminDashboard/Controllers/RoleController.cs
@@ -16,15 +16,15 @@ namespace AdminDashboard.Controllers
         }
         public async Task<IActionResult> Index()
         {
-            var roles =await _roleManager.Roles.ToListAsync();
+            var roles = await _roleManager.Roles.ToListAsync();
             return View(roles);
         }
         [HttpPost]
-        public async Task<IActionResult> Create(RoleFormViewModel model) 
+        public async Task<IActionResult> Create(RoleFormViewModel model)
         {
-            if(ModelState.IsValid)
+            if (ModelState.IsValid)
             {
-                var roleExists =await _roleManager.RoleExistsAsync(model.Name);
+                var roleExists = await _roleManager.RoleExistsAsync(model.Name);
                 if (!roleExists)
                 {
                     await _roleManager.CreateAsync(new IdentityRole(model.Name));
@@ -42,10 +42,17 @@ namespace AdminDashboard.Controllers
         public async Task<IActionResult> Delete(string id)
         {
             var role = await _roleManager.FindByIdAsync(id);
+            if (role != null && (role.Name == "Admin" || role.Name == "User"))
+            {
+                ModelState.AddModelError("Name", "Cannot delete this role!");
+                return View("Index", await _roleManager.Roles.ToListAsync());
+            }
+
             if (role != null)
             {
                 await _roleManager.DeleteAsync(role);
             }
+
             return RedirectToAction(nameof(Index));
         }
